requirements.txt
Django==2.2.5
djangorestframework==3.10.3
django-oauth-toolkit==1.2.0
drf-yasg==1.16.1
